Refactoring: extract to internal module bpf_trace (bpf-profile #9)
